cli: Fix configure failure on first use
Commit 54926c4 exposed a bug in "cfncluster configure" which
caused configure to fail if the base directory for the config
file didn't exit (in the default case, ~/.cfncluster/).  This
wasn't a common problem prior to 54926c4, because logging was
saved to ~/.cfncluster/, and it *did* create the directory
properly.  So create the directory for the config file basename
and avoid the issue, regardless of logging behavior.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>