Query cloudformation for template when starting cluster
The start command does not allow a user to provide a custom template. Without
that, the template selection logic will fallback to using the default template
setup in the global section of the config file, causing the ASG to adopt
unintended scaling values. This change makes it query CloudFormation instead, as
in the case of a cluster update.

Fixes #325.

Signed-off-by: Raghu Raja <craghun@amazon.com>