Fix return value check for cfncluster-release-check script
run_test() doesn't return a value, so retval is None.
None means "no value" so the test retval == 0 can give erratic behaviour.

We are going to ignore the return value of run_test(), because there is no one
and set retval to 0 after the run if no Exception are raised.

Signed-off-by: Enrico Usai <usai@amazon.com>