Fix Torque node allocation + job sleep tuning
To fully allocate a job on one node we need to get the nomber of
processor from the node list.

Concerning job sleep, the sum of the two jbs sleep time should exceed
the timeout, so that if they would run on the same instance the whole
test would fail. At the same time we cannot simply split the full
timeout by 2 because the intance takes from 3 to 6 minutes to come up.
So we need to absorb this time during the computation of the first job.