Improve job slots retrieval for SGE and Torque
It may happen computing nodes are not yet available or free when the
jobs are ready to be submitted. This change waits 15 secs for the
computing nodes to be available to get the number of slots available.
If there are no nodes available the script fails with an error message.