Improve cfncluster-release-check integration test
cfncluster-release-check integration test runs for few hours spawning
tens of clusters across the supported regions.
Interrupting this process means to leave a number of clusters up and
running.

This patch wants to gracefully manage the interrupted termination,
cleaning all the clusters that have been created.
This is obtained with the following changes
1. Capture the terminate/interrupt signals and manage the termination of
the running processes. This will raise an exception in all those threads
where the main operations are running and prevents further commands to
be executed. The finally block then copes with the removal of the
created clusters.
2. The name of the cluster has been made reasonably unique by adding
the timestamp of the launch. In this way we have not to wait all the
clusters to be deleted and we can launch a second instance.
3. The delete cluster operation is now not suspensive.

Changed top level process mgt and signal handling

 * tests/cfncluster-release-check.py
It has been refactored in order to split execution into a parent + child
processes.
Child process is the actual tests runner. Since python propagates SIGINT
signal to threads and sub-processes in an unpredictable and unmanageable
way, SIGINT is ignored by the child process that manages instead the
SIGTERM/SIGHUP signals.
Upon receiving a SIGTERM signal all the processes are killed and all
clusters deleted.
Due to racing conditions when trying to delete a cluster that is just at
the  very early stages of its creation, 'cluster delete' has been made
more robust by retrying a number of times when the system recognizes the
'cluster create' was interrupted.
The main cycle that waits for all the worker threads to complete was
changed in order to avoid the join() call that would prevent the
application from correctly handling signals.
The child process also copes with situations where the parent process
has been killed by a SIGKILL and in this case the child suicides.

Parent process is aimed to capture the process termination signals and
to send a SIGTERM to the child. In any case the parent process waits for
the child to terminate.

Signed-off-by: Maurizio Melato <mmelato@amazon.com>