Clean the directory in which vendored cookbooks will be placed
Clean the directory in which vendored cookbooks will be placed prior to
execute the vendor command.
This solve the problem when a cookbook dependency needs to be updated,
e.g. when Chef version is upgraded.
Without cleaning up the directory, which contains the cookbook cache
store built during the AMI creation, the old version of the cookbook
dependency will be used, potentially breaking the new requirements
of the new Chef version.

see https://github.com/berkshelf/berkshelf/issues/381 for details.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>