Improve integration tests for the scaling process
Add to the integration tests defined in cfncluster-release-check
a more granular test to verify that scale-up and scale-down features
work correctly. The integration tests are now able to verify that
the cluster is correctly scaled-up and that all created nodes are
available to the scheduler. It is also possible to customize the
scaling range of the cluster and make sure that, when some of the
compute nodes are configured to be retained in absence of jobs, this
condition is verified.
The change also refactors the integration tests structure by moving all
tests defined in cluster-check bash script into the
cfncluster-release-check python script. This was done in order to
increase the cohesion and future extensibility of integration tests.

Signed-off-by: Francesco De Martino <fdm@amazon.com>