Generate hostfile on Batch's main node and mount a volume using nfs on the compute nodes
generate_hostfile script takes as input the shared directory and the destination directory where the hostfile should reside
It assumes and checks for the existence of some of the awsbatch environment variables.

mount_nfs script takes as input the master ip address and the directory that needs to be shared. It assumes and checks for the existence of some of the awsbatch environment variables.

Signed-off-by: Balaji Sridharan <fnubalaj@amazon.com>