Implement docker images build for awsbatch scheduler
Docker images needed to execute AWSBatch jobs are automatically built
and uploaded to a stack specific ECR repository.
No additional dependencies are required to run cfncluster and life-cycle
of all created resources is managed by the cfn batch substack template.
An S3 bucket, where resources required to build docker containers images
are stored, is automatically created right before triggering the
stack create operation. This bucket is then deleted by a CloudFromation
CustomResource triggered on stack deletion.
Here is the list of steps that are executed to build docker images:
1. An S3 bucket is created and all resources contained in resources/batch
dir are uploaded to it. The bucket will contain 2 dirs: custom_resources_code
and docker. custom_resources_code stores all the scripts executed by
Lambda functions defined in the CloudFormation template; docker dir contains
the definition of the docker images to build and the CodeBuild specifications
and scripts needed to create the containers.
2. The creation of the cfncluster main template is triggered. cfncluster.cfn
then triggers the creation of the batch.cfn substack.
3. An ECR repo is created and its name is stored as stack output.
4. A CodeBuild project that builds docker images is created. This project,
when triggered, creates all docker images and uploads them to the ECR repo.
5. A CloudWatch Events Rule is created in order to catch all CodeBuild
events that are triggered when a build reaches the SUCCEEDED, FAILED or STOPPED
status. This rule has a Lambda as a target, whose role is to notify CloudFormation
that the docker images build is completed.
6. A CustomResource backed by Lambda is started. This lambda takes care
of triggering the CodeBuild project on stack creation and of removing all
images from the ECR repo on cluster deletion.
7. The stack enters a Wait state (defined as a WaitCondition).
8. The lambda triggered on CodeBuild completion, notifies the stack
(by calling the WaitHandle). The stack finally completes the creation.

In order to build additional docker images, simply create a new directory
under docker folder and add the Dockerfile definition for it. All images
are automatically uploaded to the ECR repo created by the CloudFormation
template and are tagged with a name that matches the parent directory where
the Dockerfile is contained.

The S3 bucket, created by the cli and containing all batch resources, is
deleted by CloudFormation on stack deletion. This is done in order to
remove the bucket also when the user kills the cli process during stack creation.
Additionally to reduce to the minimum the risk of not deleting this bucket,
the CustomResource which takes care of the deletion is defined in the main
cfncluster.cfn stack and is created among the first resources.

Signed-off-by: Francesco De Martino <fdm@amazon.com>