Disable Ganglia by default
The default in the AMI recipes has changed, now Ganglia is disabled by
default. Now th user has to explicitly pass ganglia_enabled = 'yes' into
the ExtraJson parameter.

Output in the CloudFormation template was kept but it's now specified
the feature is disabled by default.

Port 80 is also removed from the SecurityGroup of the Master node.

Signed-off-by: Maurizio Melato <mmelato@amazon.com>