Fix Tox Warning
Fixes:
```
W1001 Ref to resource "RootRole" that many not be available when when condition "UseS3ReadPolicy" is True and when condition "CreateEC2IAMRole" is False at Resources/S3ReadRolePolicies/Properties/Roles/0/Ref

W1001 Ref to resource "RootRole" that many not be available when when condition "UseS3ReadWritePolicy" is True and when condition "CreateEC2IAMRole" is False at Resources/S3ReadWriteRolePolicies/Properties/Roles/0/Ref
```

I changed the `UseS3Read[Write]Policy` condtion so it checks if the `CreateEC2IAMRole` is true before creating the S3 Policices. This ensures it'll never reference a non-existant RootRole.

Signed-off-by: Sean Smith <seaam@amazon.com>