Check AWS Account limits before cluster creation
At cluster creation/update/configuration time the AWS account capacity
will be verified by trying to instantiate a number of instances
equal to the configured max_queue_size.

I added the ssm:GetParametersByPath policy to the documentation, since
it is required to get the latest amazon linux ami id to use for the test.

In the awsbatch case we are calculating the number of instances for the
test according to the instance_type, the number of vcpus for the instance_type
and the selected max_vcpus parameter.
The test will be skipped if the compute instance type is set to "optimal".

Signed-off-by: Enrico Usai <usai@amazon.com>