Fixed parsing issue when using ec2_iam_role option
Fixed issue with parsing error when trying to use a custom IAM role with the ec2_iam_role option:

"Unexpected error of type ValueError: too many values to unpack (expected 2)"

The code is looking for two values for this option, but there should only be one (the name of the IAM role). This change has been tested in my environment, and has resolved my issue.