Fix and uniform logging usuage in pcluster cli
The Python logging module organizes loggers in a hierarchy. We were configuring
the logger for "pcluster.pcluster" and then use logging.getLogger(__name__) in
some of the modules. This was causing messages to use the default root logger,
hence the output was not being displayed. I uniformed this behaviour by
configuring the logger on "pcluster" and then using it with
logging.getLogger("pcluster." + __name__). As an additional future enhancement
it would be nice to configure the logger by using a configuration file so that
an expert user can customize its behaviour if necessary.

Signed-off-by: Francesco De Martino <fdm@amazon.com>