Merge cloudwatch-logging feature branch
Create cloudwatch log group via cfn template.

Implement CloudWatch logging integration tests.

These tests verify:
* whether or not the log files that were expected to exist on the master
and compute instances actually do exist
* whether or not a log group was created depending on whether the
feature was enabled
* whether or not the log streams for non-empty logs on all of the
instances were created as expected
* whether or not these same log streams contain data
* whether or not these same log streams contain the last line read from
the logs that they represent

There are known issues with these tests:
* They haven't yet been tested with awsbatch as the scheduler
qualified domain name is used where the local domain name is expected
* The tests do not verify that the log groups are deleted when the
cluster is deleted
* The tests do not verify that the correctness of the timestamp formats
used to set the event times for log events.

Add script to get archive of cluster's log group

This is the initial check-in for a script that enables a user to extract
all of the logs in a given cluster's CloudWatch log group and create a
tarball of the extracted data. This allows for ease in attaching logs to
tickets.

Potentially useful enhancements include:
* Allow for optionally deleting data from the s3 bucket that serves as
the intermediary between CloudWatch and the local filesystem.
* Enable a user to specify that the data only be exported to s3 and not
downloaded to the local filesystem.

Signed-off-by: Tim Lane <tilne@amazon.com>