Create pcluster dcv command and enable dcv settings in config file
We are adding a new command to the pcluster CLI that permits to connect
through NICE DCV to the master instance of a running cluster.

From now the [cluster ..] section accepts the dcv_settings parameter
referring to a [dcv ..] section.
To enable the DCV feature the user need to explicitly set the "enable"
parameter to "master".

Note: I extended the ParallelCluster instance policy to add s3 bucket
requested to verify NICE DCV license

This patch is strictly related to the one adding recipes for DCV on Centos7
in the ParallelCluster coookbook package.

+ Add unit tests for configuration parsing
+ Add integration tests for DCV recipes
+ Refactor ssh command by defining a function to retrieve master and user
+ Define global variable for pcluster issues link and use for dcv error

Signed-off-by: Enrico Usai <usai@amazon.com>