Fix to_file bug for BoolParam with default=True.
In Param.to_file, there is a conditional that is meant to:
1. check whether the parameter has a value, and
2. check whether the parameter is equal to its default.

When boolean parameters that default to True have a value of False, they
should be written back to the file, but the current logic does not allow
for that.

To fix that: explicitly compare the parameter's value to None when
checking if it has a value. Add a unit test to verify the fix.

Signed-off-by: Tim Lane <tilne@amazon.com>