Add script to get archive of cluster's log group
This is the initial check-in for a script that enables a user to extract
all of the logs in a given cluster's CloudWatch log group and create a
tarball of the extracted data. This allows for ease in attaching logs to
tickets.

Potentially useful enhancements include:
* Allow for optionally deleting data from the s3 bucket that serves as
the intermediary between CloudWatch and the local filesystem.
* Enable a user to specify that the data only be exported to s3 and not
downloaded to the local filesystem.

Signed-off-by: Tim Lane <tilne@amazon.com>