Improve EC2 instance types validator to get the list from pricing file
The boto3 metadata are not updated, it is required to upgrade botocore
library to see the newest instances.
The pricing file is a list specific for the product.

From now the compute_instance_type_validator automatically calls
ec2_instance_type_validator if the scheduler is not awsbatch.

I'm using two different methods to retrieve compute and master instance types
because compute instance type list must contain also families and
"optimal" keyword in case of awsbatch scheduler.

Signed-off-by: Enrico Usai <usai@amazon.com>