Make pcluster update command safe
This commit introduces a new workflow for the pcluster update command
which makes use the new class ConfigPatch to check the differences
between the existing cluster configuration and the new one wanted by the
user. The update process is allowed only if the found changes are
marked as safe by the checking mechanism.
All the detected changes are displayed to the user to clarify what is
happening behind the scenes.

An Update Policy has been assigned to each configuration parameter.

New integration tests have been introduced for the pcluster update
command to check (possibly) all the update policies assigned to
parameters in mappings.py.

Signed-off-by: ddeidda <ddeidda@amazon.com>