Handle feature-specific logs in CW integ tests
[This cookbook PR](https://github.com/aws/aws-parallelcluster-cookbook/pull/442) adds DCV logs to the config file used to generate the
CloudWatch agent configuration. This same config file is used by the
CloudWatch logs integration tests to generate the list of logs which are
expected to (a) exist on the various nodes in a cluster and (b) be
represented in the log streams present in a cluster's log group (if the
log file on the node is not empty).

This PR ensures that feature-specific logs do not result in test
failures if those logs do not exist. The reason for this behavior is to
avoid duplicating functionality that should be present in the
integration tests specific to the same feature as the logs.

As an example: some of the DCV logs for which support was added are not
created until `pcluster dcv connect` is called to start a session. It's
preferable to keep the logic that exercises this in the DCV integration
tests.

The CloudWatch logs integration tests will be modified to allow other
integration tests to verify that the logs they expect to be collected
are, but that can be added later.

Note that feature-specific logs are still validated to an extent. If
they are present and not empty, their log streams are validated per
usual.

Signed-off-by: Tim Lane <tilne@amazon.com>