Enable validation of dockerfile entry point deps
This PR enables more thorough validation of the dependencies (and
functionality) of the docker image generated when a cluster uses
awsbatch as its scheduler. It does this via two changes:
1. The `batch_dockerfile_deps` mark has been added to the awsbatch
scheduler test function, in order to allow for selecting only this test.
2. An EFS filesystem is configured in the awsbatch scheduler test
function's config file, in order to exercies the `mount_efs.sh` entry
point script in the generated docker image.

Signed-off-by: Tim Lane <tilne@amazon.com>