Keep all of a cluster's log groups after deletion
The previous commit implemented and tested a flag that allowed the
CloudWatch log group configured by the user to persist past cluster
deletion. This commit extends the actions taken when the `--keep-logs`
flag is passed to the `plcuster delete` command such that all of a
cluster's log groups (even those not configured by the user, like those
for the cleanup resources action or AWS Batch substacks) are kept after
cluster deletion.

Signed-off-by: Tim Lane <tilne@amazon.com>