Refresh PClusterConfig instance when a Section label is updated
This commit adds a refresh() method to PClusterConfig to allow marking
an instance as stale, so that next calls to object methods trigger a
refresh() operation. This new method is now called from Section.label
property, so that updating a session's label will trigger a refresh of
its parent PClusterConfig, making the sections dict() to be updated
accordingly. A new test has been added to check the correct behaviour in
this case.

Signed-off-by: ddeidda <ddeidda@amazon.com>