Pass shared dir path to CloudWatch logs test class
Enable passing the path to the shared directory configured for a cluster
to the feature-specific CloudWatch logging tests. This is required
because the CloudWatch logging tests get output from commands run on the
compute nodes by creating a temporary directory in the cluster's shared
directory and re-directing all output to files in that directory.

These changes were tested by running the affected integration tests locally.

Signed-off-by: Tim Lane <tilne@amazon.com>