Only pin specific PyYAML for python 3.4
This commit conditionally specifies the PyYAML version installed as a
ParallelCluster dependency based on the version of python used.

PyYAML 5.3 is not compatible with python 3.4, so for this version of
python, pin to 5.2. Otherwise, use the latest compatible version.

`requirements34.txt` was added for users manually installing
dependencies with one of the affected versions of python.

Signed-off-by: Tim Lane <tilne@amazon.com>