Use sanity check logic to handle errors from get_latest_alinux_ami_id
At the moment, the latest alinux ami is used to check the account capacity.
Before this patch, if there was an error retrieving the latest alinux ami, the cluster creation was aborted.
With this patch, the error is blocking or non-blocking depending on the value of the sanity check parameter.

Out of the scope of this patch is to use a ParallelCluster AMI or a custom ami to do the capacity check.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>