Disable EBS tests in China
The EBS tests in China timeout regularly when using AWS batch as the
scheduler. This has been remedied in develop by increasing the timeout,
but this change has not been applied to the release branch or (more
importantly) the most recent version of aws-parallelcluster 2.5 in PyPI.
Thus, this commit disables the EBS tests in China.

Signed-off-by: Tim Lane <tilne@amazon.com>