Fix bucket creation utility function
From now the function will raise the ClientError exception in each case
and it is up to the caller to manage the different cases.

The function is only used to create the internal s3 bucket used for
the awsbatch scheduler.

Since the awsbatch bucket is removed a cluster deletion time I removed
the condition to skip the BucketAlreadyOwnedByYou exception,
because we need to be sure that the bucket belongs to a specific cluster
and is not a bucket previously created, already owned by the user.

Then I split the try blocks for the cluster creation and the resource
artifacts upload, because it can happen that the bucket creation fails
and the code in the catch block was trying to delete a nonexistent bucket.

I also changed the log level to CRITICAL.

Refactor batch bucket creation function to be more testable:
- Moved bucket name generation in a separate function
  generate_random_bucket_name.
- Moved resources folder discovery into the
  create_bucket_with_batch_resources function.

Added unit test for:
- generate_random_bucket_name
- create_s3_bucket
- create_bucket_with_batch_resources

Signed-off-by: Enrico Usai <usai@amazon.com>