Fix missing region parameter in test_spot_default
After the new pytest release (5.4.0) the test was failing due to the
missing declaration of the "region" parameter.
From the changelog:

5712: Now all arguments to @pytest.mark.parametrize need to be
explicitly declared in the function signature or via indirect.
Previously it was possible to omit an argument if a fixture with
the same name existed, which was just an accident of implementation
and was not meant to be a part of the API.

https://docs.pytest.org/en/latest/changelog.html
Signed-off-by: ddeidda <ddeidda@amazon.com>