Add Arch CFN parameter required for ARM support
This commit adds the Arch CFN parameter to the configuration code.
This new parameter is a derived entirely from other parameters in the
configuration file, and is thus not permitted to be configured there.

Also included are validators that cause cluster creation to fail fast
when the architecture-specific components are not compatible with each
other or with the features enabled by the configuration file. The
specific validators are:
* Ensure the master and compute instance types have compatible instance
  architectures
* Ensure a custom AMI supports the architecture implied by the instance
  types
* Ensure that ARM instances are used only on OSs for which ARM base
  AMIs are available
* Ensure that incompatible features like disabling hyperthreading, FSx,
  or the Intel HPC platform are not enabled when using ARM instance
  types

Among the many things still TODO:
* adding more tests for the CLI code added
* ensuring all of the EC2 API calls made in the unit tests are mocked
* making the required changes to the CFN template