Add Arch parameter to CFN template
This commit adds the Arch parameter to the CFN template and modifies
AMI selection such that a architecture-specific mapping is used
depending on the value of this new parameter.

NOTE: the AMI IDs in the ARM map are from my personal repo. This will
be changed before checking it into the main repo.