Change AWS account limit check function to be a configuration tester
When executing dry-run the run_instance function is not allocating
the required number of nodes so the test by using max size was useless.

From now the code is no longer able to check AWS account limits
for the instance types but only verify if the provided configuration is valid.

Tested cluster creation.

Signed-off-by: Enrico Usai <usai@amazon.com>