Change AWS account limit check function to be a configuration tester
When executing dry-run the run_instance function is not allocating
the required number of nodes so the test by using max size was useless.

From now the code is no longer able to check AWS account limits
for the instance types but only verify the provided configuration is valid.

Tested cluster creation.

Error codes description here:
https://docs.aws.amazon.com/AWSEC2/latest/APIReference/errors-overview.html

Signed-off-by: Enrico Usai <usai@amazon.com>