Retrieve MasterIP from stack output when stack status is valid
When Cloudformation stack is in a valid status, the master IP must be retrieved from the stack output.
In this way, the selection to whether to take the public IP or the private IP is left on what is reported by the stack output.

This solves an issue with the combination
use_public_ip = false in the cluster configuration and Auto-assign-ipv4 = true in the master subnet configuration:
in this case the Cloudformation stack output is reporting only the private Master IP, but the IP taken and used by the command `pcluster ssh` and `pcluster dcv connect` was the public Master IP.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>