Decrease scaledown_idletime in slurm GPU test
Test was failing sporadically because the time interval between the
first job submitted by the test exiting and the last compute node
scaling down would be greater than the value of MinJobAge in the default
slurm.conf. This means that when the test went to verify that the jobs
it submitted in order to get the cluster to scale up submitted
successfully, slurm had (sometimes) already purged the job's history.

Decreasing scaledown_idletime improves the chances of the job statuses
being queried before the history is purged.

Using a greater value for MinJobAge in slurm.conf would be a more robust
way of handling this issue, but this approach is being tried first
because the parameter is directly configurable in the test whereas
changing MinJobAge would require modifying the config file in place and
restarting slurmctld.

Signed-off-by: Tim Lane <tilne@amazon.com>