Handle test metrics `feature` dimension in pytest
Prior to this commit, we added the `feature` dimension to the JSON
reports used to generate metrics for our integration tests as a
post-processing step. This led to situations where the `feature`
dimension was present in reports where none of the other dimensions
were.

To remedy this, add a `feature` user_property to test items via a pytest
hook, which then allows for metrics to be generated for this dimension
in the same way it is for the others (OS, region, scheduler, instance).

Signed-off-by: Tim Lane <tilne@amazon.com>