Generate amis.txt based on updated CFN template
Previously amis.txt was being generated based on the updated AMI
mappings passed to the function that wrote the file. This could result
in some regions not being contained in the file if they were not
updated.

Change the function such that amis.txt reflects the most recent changes
to the CFN template.

Signed-off-by: Tim Lane <tilne@amazon.com>