Prevent writing architecture parameter to config
Previously the architecture parameter was being written back to the
config files produced by `pcluster configure` when the original
`master_instance_type` supported a different architecture than the one
supported by the `master_instance_type` provided via the interactive
prompt. This change prevents that from occurring.

The same is done for the cluster_config_metadata parameter, although it
was not as critical in this case because its presence in a config file
does not cause it to fail validation.

A test case was added to the corresponding unit tests. To ensure the
test case fails when the fix is not in place, an extra check was added
to ensure that only the expected configuration keys are in the resulting
config file.

Signed-off-by: Tim Lane <tilne@amazon.com>