Add CloudWatch dashboard for head node EC2 metrics
When creating a cluster, a CloudWatch dashboard will be automatically created.
It contains information about the Head Node: metrics about the EC2 instance, the EBS/RAID volumes and EFS/FSx file systems.

The code is flexible: it allows to add `extra_parameters` to properties of each widget (such as the title, as is already done).

## Tests

* Added test to check the jinja template rendering
* Added test to check that the function is called when calling `_create_bucket_with_resources`.
* Manual tests done with sge, torque and awsbatch but most of the tests were done with slurm.
* If changing the code (e.g. by adding extra_parameters) some problems might happen when testing: if a sentence in the rendered template is too long, it will break the `rendering_template` test.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>