Create automatic Master CloudWatch dashboard with unconditional metrics
Previously, no CloudWatch dashboard was created when launching a cluster. Now, a dashboard is created in the CloudFormation stack. This dashboard contains graphs about the Master's metrics. Those metrics concern the EC2, EBS, RAID, EFS and FSx products. The provided graphs are only about the metrics that are always present when using one of those solutions (for example, the EBS BurstBalance metric depends on the volume type and is thus not yet included).

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>