Don't poll until automatic backup creation starts
Previously the FSx automatic backup test did not wait until the
scheduled automatic backup creation time before it started polling for
the backup. This resulted in sporadic failures because a large portion
of the alloted polling time would be spent before the automatic backup
creation could possibly start.

To fix this, the following changes were made:
* The function that ensures the automatic backup is created as expected
  starts by sleeping until the scheduled automatic backup start time.
* After sleeping until the scheduled start time, the same function then
  waits up to five minutes for the automatic backup creation to start.
* Once it starts, it then waits for up to seven minutes for the
  automatic backup to become available.

Signed-off-by: Tim Lane <tilne@amazon.com>