Restore default values for disable_hyperthreading and enable_efa
This commit fixes a regression introduced in https://github.com/aws/aws-parallelcluster/pull/1928, which made these parameters mandatory either in the cluster or in queue sections.

The new behavior allows configuration files to omit these parameter, with the meaning of setting them to false.
The validation checks that the parameters are specified only in cluster or in queue section, or that at least their valuea are the same in both sections.

Signed-off-by: ddeidda <ddeidda@amazon.com>