Add Route53 hosted zone and Lambda cleanup function
A new private hosted zone is created within the cluster.
The domain name is: <clustername>.parallelcluster.

We extended the cleanup Lambda function to delete DNS record
because the hosted zone cannot be deleted if there are DNS record on it.

Right now the DNS domain name is passed as input of the HITSubstack
and propagated as output.
Both HostedZone Id and domain name are output of the substack.
The hosted zone is required by node daemons to add record sets for new nodes
and by Lambda function to delete the record set before cluster deletion.

Other changes:
+ Added policy to permit secondary index queries in dynamodb and permissions for route53 updates
+ Added condition for HIT related policies: DescribeInstances and FleetTerminatePolicy

Tests:
+ Cluster creation
+ Cluster deletion with 1 record set
+ Cluster deletion without record sets
+ Dynamodb query from head node

Signed-off-by: Enrico Usai <usai@amazon.com>