Split Lambda function to cleanup Route53 and S3/EC2 resources
The lambda to cleanup the S3 Bucket must be executed as soon as possible
instead the lambda functions to terminate the instances and to cleanup
the route53 must be executed after the master is stopped to avoid
other instances or record set to be added after the Lambda execution.

Moved the Lambda related code and the policy specific to HIT in the
compute fleet hit substack.

# Implementation details
* The trick to have a dependency in the Master substack is to add
  the TerminateComputeFleetCustomResource and the CleanupRoute53CustomResource
  as custom parameter "DependsOnCustomResources".
* To have a valid substack when Route53 resource is not created we
  are using input parameters as metadata.
* We are creating a policy in the substack and attaching it to the RootRole
  if the customer is not using a custom EC2 IAM Role

Signed-off-by: Enrico Usai <usai@amazon.com>