Stop using S3 configuration files
* Use DescribeInstanceTypes and DescribeInstanceTypeOfferings to
validate instance types and get vCPU info.
* For validating compute instance types to be used with Batch, work
around the lack of an API by attmepting to create a compute
environment using a nonexistent instance type, and then parse the
supported instance types for the given region from the error
message. If this information cannot be obtained in this way, fall back
to using the list of available instance types in the region.