Add Cluster name and queue_settings validators and fix file_path regex
Before this commit, the cluster name and queue_settings were not validated. Now, those are verified to ensure that they consist only of alphanumeric characters and hyphens (but begin only with letters) and that the lengths of the cluster name and of the each queue name in the queue_settings are respectively less than 60 and 30. Testing for those modifications has also been added.

The limitation on the cluster name length has been chosen based on two observed limitations:
- The LaunchTemplateName in compute-fleet-hit.substack.cfn.yaml has name ${ClusterName}-{{ queue }}-{{ instance_type }} and the LaunchTemplateName can have maximum 128 characters. Choosing 30 for the queue name, 21 for the instance type name, we still have 75 chars left for the queue name;
- The more limiting factor comes with HostedZone names which consist of a series of labels separated by dots. Each label can be up to 63 bytes long. The length of the cluster name can thus be at most 63 chars (as we do not want to introduce dots, which are subdomains in DNS).

Limitation on queue name: The hostname is queuename-st|dy-instancetype-N with N being a number and this hostname should be less than 64 chars long. That is why a length of 30 chars was chosen for the queue name.

Resources section in the main stack and in batch substack have been changed as the ClusterName/AWS::StackName gets truncated by CloudFormation (max 100 characters when concatenating (using hyphens) the main stack name with substack names with 13 random digits). The most limiting substack name is ComputeFleetHITSubstack.

The file path regex has also been fixed. Testing about this regex has also been improved (modified and created some).

A little typo has been fixed in the cloudformation README to provide the real command to render troposphere templates.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>